requirements.txt
streamlit
sentence-transformers
faiss-cpu
pymupdf
whisper
transformers
pydub
python-dotenv